feat(features,signals): three-layer Phase-2 modules -- horizon aggregator + signal engine + regime gate + micro-SM extension
Implements the modules-only slice of Phase 2 of the v0.2 three-layer
architecture refactor (see design_docs/three_layer_architecture.md
sections 7.5, 7.6, 8 for the aggregator / horizon signal engine /
micro-state-machine extension respectively).

This PR ships the new feature- and signal-layer modules with their
unit, property, and contract tests, but does NOT yet wire them into
``feelies.bootstrap`` or ``feelies.kernel.orchestrator``.  The wiring
PR (Phase 3) lands those changes once the alpha-loader 1.1 dispatch,
``signal_layer_module``, and ``portfolio_layer_module`` exist for
the orchestrator to compose against.  Until then the new modules are
fully isolated -- import-time only -- so the legacy execution path
remains bit-identical (Inv-A in the implementation plan).

Why
---
Phase 1 (PR #5) shipped the Layer-1 (Sensor) and Layer-2-protocol
contracts.  Phase 2 supplies the *engines* that consume sensor
output, build per-horizon feature snapshots, and translate those
snapshots into Signal events under the closed v0.3 TrendMechanism
taxonomy:

  Sensor -> [HorizonAggregator] -> HorizonFeatureSnapshot
         -> [HorizonSignalEngine] -> Signal
                                  ^
                                  +-- RegimeGateDSL gate

The micro state machine is extended with horizon-gated transitions
so signal emission can flow through the same deterministic SM the
legacy path uses (Inv-5 -- replay parity preserved).

What lands here
---------------
* feelies.features.aggregator (HorizonAggregator)
    - Per-symbol, per-horizon snapshot builder.  Maintains lock-step
      evaluation order against the HorizonScheduler ticks shipped
      in PR #5.  Emits HorizonFeatureSnapshot events with stable
      composite keys (symbol, horizon_seconds, snapshot_seq).
* feelies.features.protocol (HorizonFeature)
    - Protocol every per-horizon feature must satisfy.  Mirrors the
      Sensor protocol from PR #5 so the aggregator can treat sensor
      readings and derived features uniformly.
* feelies.features.legacy_shim (LegacyFeatureShim)
    - Adapter that exposes legacy per-tick feature definitions
      through the HorizonFeature protocol.  Lets LEGACY_SIGNAL
      alphas keep their feature_definitions block during the
      transition window before workstream D.2 retires the layer.
* feelies.signals.horizon_engine (HorizonSignalEngine,
  RegisteredSignal)
    - Subscribes to HorizonFeatureSnapshot, dispatches to every
      registered SIGNAL alpha for the matching horizon, applies the
      RegimeGate (if declared), and emits Signal events tagged with
      the closed v0.3 TrendMechanism enum.
* feelies.alpha.cost_arithmetic (CostArithmetic)
    - Stdlib-only leaf utility used by HorizonSignalEngine to fold
      transaction-cost adjustments into signal payloads (gate G16
      "cost arithmetic completeness" -- enforced at the loader in
      Phase 3).  No internal feelies imports; included here because
      horizon_engine cannot resolve without it.
* feelies.signals.regime_gate (RegimeGateDSL)
    - Pure-Python evaluator for the regime-gate DSL declared in
      schema 1.1.  Deterministic, side-effect free, no I/O; gate
      decisions are folded into the signal hash (Inv-5 layer-2 hash).
* feelies.kernel.micro extended (no removals)
    - Adds a horizon-gated transition table alongside the existing
      tick-driven SM.  Default config keeps the legacy path live;
      the extended SM only activates when an alpha declares a
      horizon_seconds member of platform.horizons_seconds (gate G7).
* feelies.features small library type-tightening
    - Wraps EWMAComputation, RollingVarianceComputation, and
      ZScoreComputation return values in float() to silence mypy
      strict on the new aggregator import path.  Pure cosmetic;
      no numerical behaviour change.

Tests
-----
* tests/features/test_aggregator.py
    - Lock-step ordering against scheduler ticks; stable composite
      key; multi-symbol multi-horizon fan-out; warmup gating.
* tests/features/test_horizon_feature_protocol.py
    - Protocol smoke; type-checking surface; stale-state rejection.
* tests/features/test_legacy_shim.py
    - Round-trip of legacy FeatureDefinition through the shim;
      deprecation-warning emission; bit-identical numerical output.
* tests/signals/test_horizon_signal_engine.py
    - Per-horizon dispatch; TrendMechanism enum enforcement; signal
      hash determinism; multi-alpha conflict resolution.
* tests/signals/test_regime_gate_dsl.py +
  tests/signals/test_regime_gate_dsl_props.py
    - DSL parser, evaluator, and Hypothesis property tests covering
      the closed RegimeState taxonomy.
* tests/kernel/test_micro_extended.py
    - Extended SM transition table; legacy path unchanged when the
      horizon_seconds field is absent.
* tests/kernel/test_micro_sm_signal_gate.py +
  tests/kernel/test_micro_sm_signal_props.py
    - Horizon-gated emission contract; idempotent replay; property
      tests for the SM under random tick sequences.
* tests/monitoring/test_sensor_metrics.py
    - End-to-end sensor -> aggregator -> metric collector path
      using only Phase-2 modules (no bootstrap dependency).
* tests/alpha/test_cost_arithmetic_gate.py
    - Unit coverage for CostArithmetic helpers; ships with the
      module so the leaf utility is fully exercised in this PR.

Backward compatibility
----------------------
* No legacy code path is removed.  The new modules are not yet
  wired into bootstrap or orchestrator -- they only become live in
  PR #6 (or wherever Phase-3 lands).
* features/__init__.py adds three additive exports
  (HorizonAggregator, HorizonFeature, LegacyFeatureShim).  Existing
  symbols (FeatureComputation, FeatureDefinition, WarmUpSpec)
  remain re-exported with the same public names.
* features/library.py changes are pure float() casts; numerical
  output is bit-identical.
* kernel/micro.py is extended in place.  The extended transition
  table is gated on the alpha declaring horizon_seconds; LEGACY
  alphas (no horizon_seconds) take the original path unchanged.

Stack position
--------------
Stacked on PR #5 (feature/sensor-layer-contracts).  GitHub will
auto-retarget the base to ``main`` once PR #5 merges.

Follow-ups (out of scope here)
------------------------------
* Phase-3 -- alpha-loader schema-1.1 dispatch + signal/portfolio
  layer modules + bootstrap/orchestrator wiring + reference alphas.
* Phase-4 -- composition engine + cross-sectional tracker +
  factor neutralization.

Made-with: Cursor

+++ src/feelies/features/library.py
@@ -76,7 +76,7 @@ class EWMAComputation:
         else:
             state["ewma"] += self._alpha * (mid - state["ewma"])
         state["count"] += 1
-        return state["ewma"]
+        return float(state["ewma"])
 
 
 class RollingVarianceComputation:
@@ -112,7 +112,7 @@ class RollingVarianceComputation:
 
         state["prev_mid"] = mid
         state["count"] = count + 1
-        return state["var"]
+        return float(state["var"])
 
 
 class ZScoreComputation:
@@ -151,8 +151,8 @@ class ZScoreComputation:
             state["var"] += self._alpha * (diff * diff - state["var"])
 
         state["count"] = count + 1
-        std = math.sqrt(max(state["var"], 1e-24))
+        std = math.sqrt(max(float(state["var"]), 1e-24))
         # Clamp z-score: near-zero variance early in the session (few ticks)
         # can produce arbitrarily large values that poison downstream signals.
         _MAX_ZSCORE = 10.0
-        return max(-_MAX_ZSCORE, min(_MAX_ZSCORE, diff / std))
+        return float(max(-_MAX_ZSCORE, min(_MAX_ZSCORE, diff / std)))
